management-system-backend/sales/requirements.txt
django==5.0.2
djangorestframework==3.14.0
drf-yasg==1.21.7
psycopg2==2.9.9
pika==1.3.2
requests==2.31.0
python-dotenv
django-cors-headers